Добавил ToolBar. (поиск,настройки,о программе) изменил стандартный TextView на ListView. Приложение подстраиваю под стандартный вид заметок андройда. Буду переделывать ListView на CardView. Поиск вынести из ToolBar ниже
