+++ app/src/main/java/com/example/notesapp/MainActivity.java
@@ -1,8 +1,20 @@
 package com.example.notesapp;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
+import android.content.res.Configuration;
 import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.widget.SearchView;
+import android.widget.Toast;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -10,5 +22,36 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayShowHomeEnabled(true);
+        setTitle("");
+        FloatingActionButton fab = findViewById(R.id.addNotes);
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                FragmentManager fragmentManager = getSupportFragmentManager();
+                FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+                if(getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE){
+                    fragmentTransaction.replace(R.id.FrameLayoutNoteBody, new FragmentViewAddNewNote());
+                }
+                else{
+                    fragmentTransaction.replace(R.id.fragment_notes, new FragmentViewAddNewNote());
+                }
+                fragmentTransaction.addToBackStack(null);
+                fragmentTransaction.commit();
+                //Toast.makeText(this, "Переходим на фрагмент добавления заметки!",   Toast.LENGTH_LONG).show();
+            }
+        });
     }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Здесь определяем меню приложения (активити)
+        getMenuInflater().inflate(R.menu.main, menu);
+
+        return true;
+    }
+
+
 }
