--- app/src/main/res/layout/fragment_button_add_new_note.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".FragmentButtonAddNewNote">
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/addNotes"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="16dp"
-        android:src="@drawable/ic_my_icon"
-        tools:ignore="ContentDescription" />
-</FrameLayout>
