+++ app/src/main/res/layout/fragment_notes.xml
@@ -1,11 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".FragmentNotes"
-    android:orientation="vertical">
 
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/matchParent"
+    android:orientation="vertical"
+    tools:context=".FragmentNotes">
+
+    <ListView
+        android:id="@+id/listView1"
+        style="@style/matchParent" />
 
 </LinearLayout>
