+++ app/src/main/res/values/arrays.xml
@@ -6,8 +6,38 @@
         <item>Заметка 3</item>
         <item>Заметка 4</item>
         <item>Заметка 5</item>
+        <item>Заметка 1</item>
+        <item>Заметка 2</item>
+        <item>Заметка 3</item>
+        <item>Заметка 4</item>
+        <item>Заметка 5</item>
+        <item>Заметка 1</item>
+        <item>Заметка 2</item>
+        <item>Заметка 3</item>
+        <item>Заметка 4</item>
+        <item>Заметка 5</item>
+        <item>Заметка 1</item>
+        <item>Заметка 2</item>
+        <item>Заметка 3</item>
+        <item>Заметка 4</item>
+        <item>Заметка 5</item>
     </string-array>
     <string-array name="notesBody">
+        <item>Что то написано 1</item>
+        <item>Что то написано 2</item>
+        <item>Что то написано 3</item>
+        <item>Что то написано 4</item>
+        <item>Что то написано 5</item>
+        <item>Что то написано 1</item>
+        <item>Что то написано 2</item>
+        <item>Что то написано 3</item>
+        <item>Что то написано 4</item>
+        <item>Что то написано 5</item>
+        <item>Что то написано 1</item>
+        <item>Что то написано 2</item>
+        <item>Что то написано 3</item>
+        <item>Что то написано 4</item>
+        <item>Что то написано 5</item>
         <item>Что то написано 1</item>
         <item>Что то написано 2</item>
         <item>Что то написано 3</item>
